+++ src/game/scene/world/LocalWorld.h
@@ -51,8 +51,8 @@ public:
     unsigned short getBiome(glm::vec3 pos);
 
     //Called by the Client
-    void localSetBlock(glm::vec3 pos, unsigned int block);
-    //Called form ClientNetworkInterpreter
+    void localSetBlock(glm::ivec3 pos, unsigned int block);
+    //Called from ClientNetworkInterpreter
     void setBlock(glm::vec3 pos, unsigned int block);
 
     bool solidAt(glm::vec3 pos);
